Update README.md
FireControl gives an easy way for one to enable firewall rules.
Firewall rules that have been pre-configured can be turned off in the edge router.
run the script from a python enabled device on the local network and quickly block traffic to an the internet from a specific mac or ip address.